+++ ShapeCrawler.Tests.Unit/TestFile_014.cs
@@ -13,7 +13,7 @@ namespace ShapeCrawler.Tests.Unit
         public void Slide_Elements_Test()
         {
             // ARRANGE
-            var pre = new PresentationSc(Properties.Resources._014);
+            var pre = PresentationSc.Open(Properties.Resources._014, false);
 
             // ACT-ASSERT
             var elements = pre.Slides[2].Shapes;
@@ -23,11 +23,11 @@ namespace ShapeCrawler.Tests.Unit
         public void FontHeight_Test1()
         {
             // Arrange
-            var pre = new PresentationSc(Properties.Resources._014);
+            var pre = PresentationSc.Open(Properties.Resources._014, false);
 
             // Act
             var element = pre.Slides[3].Shapes.Single(x => x.Id == 5);
-            var fh = element.Text.Paragraphs.First().Portions.First().Font.Size;
+            var fh = element.TextBox.Paragraphs.First().Portions.First().Font.Size;
 
             // Assert
             Assert.Equal(1200, fh);
@@ -37,11 +37,11 @@ namespace ShapeCrawler.Tests.Unit
         public void FontHeight_Test2()
         {
             // Arrange
-            var pre = new PresentationSc(Properties.Resources._014);
+            var pre = PresentationSc.Open(Properties.Resources._014, false);
 
             // Act
             var element = pre.Slides[4].Shapes.Single(x => x.Id == 4);
-            var fh = element.Text.Paragraphs.First().Portions.First().Font.Size;
+            var fh = element.TextBox.Paragraphs.First().Portions.First().Font.Size;
 
             // Assert
             Assert.Equal(1200, fh);
@@ -51,11 +51,11 @@ namespace ShapeCrawler.Tests.Unit
         public void Title_FontHeight_Test()
         {
             // Arrange
-            var pre = new PresentationSc(Properties.Resources._014);
+            var pre = PresentationSc.Open(Properties.Resources._014, false);
 
             // Act
             var element = pre.Slides[5].Shapes.Single(x => x.Id == 52);
-            var fh = element.Text.Paragraphs.First().Portions.First().Font.Size;
+            var fh = element.TextBox.Paragraphs.First().Portions.First().Font.Size;
 
             // Assert
             Assert.Equal(2700, fh);