+++ ShapeCrawler/Models/ISlide.cs
@@ -1,0 +1,6 @@
+﻿namespace ShapeCrawler.SlideMaster
+{
+    public interface ISlide
+    {
+    }
+}